Charan (#14)
* 11/04/2025

* 22/04/2025 - Charan

* Delete src/main/java/com/realestate/main/security/CorsFilter.java

* 24/04/2025-charan

* 24/04/2025-charan

* 28/04/2025-charan

* 07/05/2025-charan

* 09/05/2025

* 13/05/2025-charan

* 16/05/2025-charan

+++ src/main/java/com/realestate/main/repository/AgentRepository.java
@@ -9,7 +9,7 @@ import org.springframework.stereotype.Repository;
 import com.realestate.main.entity.Agent;
 
 @Repository
-public interface AgentRepository extends JpaRepository<Agent, Integer>{
+public interface AgentRepository extends JpaRepository<Agent, Long>{
 
 	Optional<Agent> findByEmail(String email);
