+++ src/main/java/com/realestate/main/repository/CustomerRepository.java
@@ -5,11 +5,10 @@ import java.util.Optional;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
-import com.realestate.main.entity.Agency;
 import com.realestate.main.entity.Customer;
 
 @Repository
-public interface CustomerRepository extends JpaRepository<Customer, Integer>{
+public interface CustomerRepository extends JpaRepository<Customer, Long>{
 
 	Optional<Customer> findByEmail(String email);
